classification code

+++ growth/src/numerical/edgegrowth2/run_numerical_multithread.py
@@ -4,7 +4,6 @@
 import sys
 import os
 
-from importlib_metadata import distribution
 pwd = os.getcwd()
 modellingpath = pwd.rpartition("modelling")[0] + pwd.rpartition("modelling")[1] 
 modellingephemeral = '/rds/general/ephemeral/user/mo2016/ephemeral/Documents/modelling'
@@ -44,7 +43,7 @@ variant= 0
 n_species=2
 
 # Specifiy number of parameter sets in parameterset file to be loaded
-n_param_sets = 2000000
+n_param_sets = 200000
 
 
 
@@ -55,7 +54,7 @@ date = date.today().strftime('%m_%d_%Y')
 
 
 
-def numerical_check(df,circuit_n, variant = variant, n_species=n_species):
+def numerical_check(df,circuit_n, variant = variant, n_species=n_species, test = False):
     # bigger field
     # L=500; dx =1; J = int(L/dx)
     # T =3000; dt = 0.05; N = int(T/dt)
@@ -67,12 +66,13 @@ def numerical_check(df,circuit_n, variant = variant, n_species=n_species):
     # boundaryCoeff=2;rate=0.01
 
     # smaller time and smaller dt 
-    L=50; dx =1; J = int(L/dx)
-    T =500; dt = 0.005; N = int(T/dt)
+    L=100; dx =1; J = int(L/dx)
+    T =1000; dt = 0.005; N = int(T/dt)
     boundaryCoeff=2;rate=0.1
 
     filename= lambda mechanism, parID: 'circuit%s_variant%s_bc%s_%s_rate%s_ID%s_L%r_J%r_T%r_N%r'%(circuit_n,variant,boundaryCoeff, mechanism,rate,parID,L,J,T,N)
-    # T =10; dt = 0.2; N = int(T/dt)
+    if test == True:
+        T =100; dt = 0.1; N = int(T/dt)
 
     # df_index = np.unique(df.index.get_level_values(0))
     for parID,ss in df.index:
@@ -86,20 +86,21 @@ def numerical_check(df,circuit_n, variant = variant, n_species=n_species):
 
         # steadystates=par_dict['ss_list']
 
- 
+        simulateGrowth=False
         try:
-            mechanism = 'edgegrowth2'
-            U_final,U_record, U0, x_grid, reduced_t_grid, cellMatrix= cn_edgegrowth2_numba(par_dict,L,J,T,N, circuit_n, rate=rate, boundaryCoeff=boundaryCoeff, tqdm_disable=True)
+            if simulateGrowth == True:
+                mechanism = 'edgegrowth2'
+                U_final,U_record, U0, x_grid, reduced_t_grid, cellMatrix= cn_edgegrowth2_numba(par_dict,L,J,T,N, circuit_n, rate=rate, boundaryCoeff=boundaryCoeff, tqdm_disable=True)
 
-            with open(modellingpath + '/growth/out/numerical/%s/%s/simulation/2Dfinal_%s.pkl'%(circuit_n,mechanism,filename(mechanism,parIDss)), 'wb') as f:
-                pickle.dump(U_final, f)
-            
-            with open(modellingephemeral + '/growth/out/numerical/%s/%s/simulation/2Drecord_%s.pkl'%(circuit_n,mechanism,filename(mechanism,parIDss)), 'wb') as f:
-                pickle.dump(U_record, f)
-            
+                with open(modellingephemeral + '/growth/out/numerical/%s/%s/simulation/2Dfinal_%s.pkl'%(circuit_n,mechanism,filename(mechanism,parIDss)), 'wb') as f:
+                    pickle.dump(U_final, f)
+                
+                with open(modellingephemeral + '/growth/out/numerical/%s/%s/simulation/2Drecord_%s.pkl'%(circuit_n,mechanism,filename(mechanism,parIDss)), 'wb') as f:
+                    pickle.dump(U_record, f)
+                
             mechanism = 'nogrowth'
             U_final,U_record, U0, x_grid, reduced_t_grid= cn_nogrowth(par_dict,L,J,T,N, circuit_n, tqdm_disable=True)
-            with open(modellingpath + '/growth/out/numerical/%s/%s/simulation/2Dfinal_%s.pkl'%(circuit_n,mechanism,filename(mechanism,parIDss)), 'wb') as f:
+            with open(modellingephemeral + '/growth/out/numerical/%s/%s/simulation/2Dfinal_%s.pkl'%(circuit_n,mechanism,filename(mechanism,parIDss)), 'wb') as f:
                 pickle.dump(U_final, f)
 
             with open(modellingephemeral + '/growth/out/numerical/%s/%s/simulation/2Drecord_%s.pkl'%(circuit_n,mechanism,filename(mechanism,parIDss)), 'wb') as f:
@@ -137,11 +138,9 @@ with open(modellingpath + '/growth/out/analytical/lsa_dataframes/lsa_df_%s_varia
 # df= pickle.load( open(modellingpath + '/growth/out/analytical/lsa_dataframes/lsa_df_%s_variant%r_%rparametersets.pkl'%(circuit_n,variant,n_param_sets), "rb"))
 df.index.names = ['parID','ss']
 total_params=len(df)
-total_params=200000
-print(total_params)
+print(len(df))
 print('loaded')
 batch_size = int(total_params/Number_of_Threads) + 1
-df = df.iloc[0:total_params]
 print(df.head())
 batch_indices = list(range(0, len(df), batch_size))
 # Create a pool of workers
@@ -166,3 +165,5 @@ for count,start_batch_index in enumerate(batch_indices):
 finish_time = time.perf_counter()
 time_taken = finish_time-start_time
 print("Time taken: %d s" %time_taken)
+with open(modellingpath + '/growth/out/analytical/instability/instability_df_%s_variant%r_%rparametersets.pkl'%(circuit_n,variant,n_param_sets), "rb") as f:
+    df = pickle.load(f)
